Do not strip leading space after an inline tag
Inline tags (e.g., `@link`, `@linkplain`, etc.) cause a break in the
text processing so what looks like the start of a new line when text
processing resumes is actually the continuation of a line that an inline
tag interrupted.

In this case, we do not want to strip the leading space. This fixes the
issue where all generated links lost a space character before the next
word.